Заменил AddPointView и EditPointView на  PointConfiguratorView. У класса .trip-main__trip-controls удалил отступ
