+++ public/css/style.css
@@ -189,8 +189,7 @@ main {
   margin-bottom: 35px; }
 
 .trip-main__trip-controls {
-  width: 600px;
-  margin-top: 36px;}
+  width: 600px;}
 
 .trip-main__event-add-btn {
   min-width: 170px;
