+++ src/presenter/trip-events-presenter.js
@@ -1,7 +1,7 @@
 import SortView from '../view/sort-view.js';
 import { render, RenderPosition } from '../render.js';
 import TripEventsListView from '../view/trip-events-list-view.js';
-import EditPointView from '../view/edit-point-view.js';
+import PointConfuguratorView from '../view/point-configurator-view.js';
 import PointView from '../view/point-view.js';
 
 
@@ -15,7 +15,7 @@ export default class TripEventsPresenter{
   init = () => {
     render(new SortView(), this.container, RenderPosition.BEFOREEND);
     render(this.tripEventsListComponent, this.container, RenderPosition.BEFOREEND);
-    render(new EditPointView(), this.tripEventsListComponent.getElement(), RenderPosition.BEFOREEND);
+    render(new PointConfuguratorView(), this.tripEventsListComponent.getElement(), RenderPosition.BEFOREEND);
 
     for (let i = 0; i < 3; i++) {
       render(new PointView(), this.tripEventsListComponent.getElement(), RenderPosition.BEFOREEND);
